pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "lang2sql"
dynamic = ["version"]
description = "Lang2SQL - Query Generator for Data Warehouse"
readme = "README.md"
requires-python = ">=3.9"
authors = [
  { name = "ehddnr301", email = "dy95032@gmail.com" },
]
dependencies = [
  "langgraph==0.2.62",
  "datahub==0.999.1",
  "langchain==0.3.14",
  "langchain-community==0.3.14",
  "openai==1.59.8",
  "langchain-openai==0.3.0",
  "streamlit==1.41.1",
  "python-dotenv==1.0.1",
  "faiss-cpu==1.10.0",
  "transformers==4.51.2",
  "langchain-aws>=0.2.21,<0.3.0",
  "langchain-google-genai>=2.1.3,<3.0.0",
  "langchain-ollama>=0.3.2,<0.4.0",
  "langchain-huggingface>=0.1.2,<0.2.0",
  "clickhouse_driver>=0.2.9,<0.3.0",
  "plotly",
  "matplotlib",
  "ipython",
  "kaleido",
  "numpy<2.0",
  "snowflake-connector-python>=3.15.0,<4.0.0",
  "databricks-sql-connector>=4.0.3,<5.0.0",
  "oracledb>=3.1.0,<4.0.0",
  "mysql-connector-python>=9.3.0,<10.0.0",
  "duckdb>=1.2.2,<2.0.0",
  "psycopg2-binary>=2.9.10,<3.0.0",
  "psycopg[binary]>=3.2,<4.0",
  "pyodbc>=5.1.0,<6.0.0",
  "crate>=0.29.0,<1.0.0",
  "pyhive>=0.6.6,<1.0.0",
  "google-cloud-bigquery>=3.20.1,<4.0.0",
  "pgvector==0.3.6",
  "langchain-postgres==0.0.15",
  "trino>=0.329.0,<1.0.0",
  "anthropic>=0.20.0",
  "sqlalchemy>=2.0",
]

[project.scripts]
lang2sql = "cli.__init__:cli"

[project.urls]
Homepage = "https://github.com/CausalInferenceLab/Lang2SQL"
Repository = "https://github.com/CausalInferenceLab/Lang2SQL"

[tool.hatch.version]
path = "version.py"

[tool.hatch.build]
include = [
  "version.py",
  "prompt/*.md",
  "src/lang2sql/**",
  "src/lang2sql/harness/prompts/*.md",
]

[tool.hatch.build.targets.wheel]
packages = [
  "cli",
  "interface",
  "engine",
  "infra",
  "prompt",
  "utils",
  "src/lang2sql"
]

[project.optional-dependencies]
# Harness core — no extra deps needed (uses stdlib + core)
harness = []
# Tool-calling LLM providers
openai = ["openai>=1.0"]
anthropic = ["anthropic>=0.20"]
# Terminal UI
tui = ["textual>=0.80", "rich>=13.0"]
# Terminal charts
viz = ["plotext>=5.0"]
# HTML chart export
viz-export = ["plotly>=5.0"]
# All harness + TUI + viz
agent = [
  "lang2sql[openai,anthropic,tui,viz,viz-export]",
]

[tool.uv]
dev-dependencies = [
  "pre_commit==4.1.0",
  "pytest>=8.3.5",
  "pytest-asyncio>=0.23",
]
